refactor: create component for each step content
